Created a set of memory-friendly collections specifically for use in the compiler.
The goal is to provide API compatibility with the JRE versions, yet use a lot less memory.  I particularly am focusing on optimizing the case of empty and singleton collections, which is an extremely common case in the compiler, and an easy win for memory.  Having lots of nodes refer to the single-instance empty set or list takes almost no memory, a default empty ArrayList or HashSet is very expensive by comparison.

Review by: spoon

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@5149 8db76d5a-ed1c-0410-87a9-c151d255dfc7