Fix build timeouts due to triplicate use of user test.  The problem is that servlet and tools depend on user, but in a context (from test's call-subproject) in which $target has been set to "test", so that user dependency is a dependency on testing user.  Again.  And again.
In general, we use subdirectory dependencies to mean "must be built before playing with this subdir in any way," but gwt.ant doesn't know which things to build and which to play with.  Fast solution is not to use depends=..., but to explicitly gwt.ant what we want.

Review by: bobv

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@5584 8db76d5a-ed1c-0410-87a9-c151d255dfc7