JSNI ::new() now targets constructors directly.
This patch removes the extra hop through a synthetic static 'new' function that is currently used for implementing JSNI ::new() invocations.

In the 99% case, the JsInvocation is replaced with a JsNew operation on the target constructor.  In rare cases where the constructor is not immediately invoked, a tear off function is creation which performs the new op internally.

Review by: tobyr@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8366 8db76d5a-ed1c-0410-87a9-c151d255dfc7