SafeHtml Templates
The SafeHtml Templates were removed from the main SafeHtml review, but are being added again here so that widgets can start using this api, even while a newer parser is in the works. This is a first pass at templates, and the api/usage are up for review.

Review at http://gwt-code-reviews.appspot.com/793801


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8657 8db76d5a-ed1c-0410-87a9-c151d255dfc7