Add statistics to optimizers
Updates the compiler optimizers to returns statistics about each pass
instead of a simple boolean.   This could be used to analyze
the effectiveness of an individual optimizer, tune the compiler
for code size/ compile time trade offs, or predict the effectiveness
of future passes of an optimizer.

Review at http://gwt-code-reviews.appspot.com/841801


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8736 8db76d5a-ed1c-0410-87a9-c151d255dfc7