Refactoring the Showcase sample to use standards mode, and make use of LayoutPanels. The new Showcase looks different, but the features are the same. The main menu is now a CellTree backed by a TreeViewModel.  The Application class, which used to perform active layout of the entire app, has been replaced by layout panels. The ShowcaseShell uses UiBinder to control the outer layout. The buttons to switch style themes have been removed. Opening a Category in the main menu prefetches the code for the sample under the category; the old behavior was for each ContentWidget to call a static method to preload other examples in the same category.
Note that this change does not update the examples within showcase aside from couple of minor tweaks. In a later patch, we will deprecate old examples, add new examples for new GWT features, and make the examples look prettier with the rest of the app.  This patch is smaller than it looks because more of the changes to the examples are related to a refactor in ContentWidget, so they are repetitive and small.

Review at http://gwt-code-reviews.appspot.com/837801

Review by: jat@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8766 8db76d5a-ed1c-0410-87a9-c151d255dfc7