JUnit ought to just serialize exceptions.
Back in the day, Exceptions couldn't be serialized.  So we had to do hacky things to support JUnit, since we needed to transmit exceptions across the wire.

Better days are here, and Exceptions are now Serializable.  So I'm pulling out some very archaic code.  The tangible benefit is: certain exceptions that could never be deserialized before, should just work now that I was able to remove the wacky custom deserialization logic.

http://gwt-code-reviews.appspot.com/870802/show


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8805 8db76d5a-ed1c-0410-87a9-c151d255dfc7