Fixes design time support for @UiField.
It turns out that previous patch was not enough.  In new one instead
of preventing failing when no owner or field found, we disable
assigning to fields and support for event handlers at all. However we
can not disable them globally, for all UiBinder templates, we should
do this only for template under design (because other templates may be
used on it). So, I've changed a little way for choosing
DesignTimeUtils implementation.

http://gwt-code-reviews.appspot.com/834802

Review by: rjrjr@google.com


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8814 8db76d5a-ed1c-0410-87a9-c151d255dfc7