Makes EventBus available outside of the gwt package, in com.google.web.bindery.event.shared.  Retrofits the old classes as deprecated wrappers around the new ones.
The new package is not quite an identical API: we use the
excuse to remove the EventHandler and HasEventHandlers
interfaces, which serve no discernable purpose. And the new
base event class is named Event, not GwtEvent

Note that the classes used directly by GWT widgets and dom
events (GwtEvent and HandlerManager in particular) are not
deprecated.

Review at http://gwt-code-reviews.appspot.com/1394803


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10014 8db76d5a-ed1c-0410-87a9-c151d255dfc7