Introducing CustomScrollPanel, a subclass of ScrollPanel that lets users define their own scrollbars. By default, we use NativeVerticalScrollbar and NativeHorizontalScrollbar, but in the future we might introduce thin scrollbars for mobile devices. The scrollbars are semi-transparent until the user hovers them, at which time they transition to 100% opacity (on browsers that do not support CSS transitions, they immediately become opaque). In RTL, the vertical scrollbar appears on the left side (many browsers get this wrong natively).
CustomScrollPanel looks and acts almost exactly like a normal scrollable div, with one important exception.  The scrollbars in a CustomScrollPanel are overlaid on top of the content instead of adjacent to it, so when scrollbars appear, the content does not require relayout. This can be useful for layout when aligning items directly above and/or below the CustomScrollPanel, because the content will always flow to the end of the panel.  CustomScrollPanel is a pre-requisite for a CellTable with fixed headers/footers and scrollable content area.  It allows the header/footer tables to be the same width as the content table regardless of whether or not scrollbars appear.  If we tried to a regular ScrollPanel, we would need active layout to adjust the width of the header/footer tables to account for the appearance of scrollbars.

In terms of implementation details, the content is actually in a scrollable div, but the scrollbars of the div are hidden behind the outer viewport.  So, all of the normal scrolling mechanisms (click and drag text, mousewheel, page up/down) work normally. We use the ResizeLayoutPanel implementation to detect when the content is resize, and show/hide scrollbars accordingly.

You can view a demo here: http://cscroll.jlabanca-testing.appspot.com/

Review at http://gwt-code-reviews.appspot.com/1427804


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10067 8db76d5a-ed1c-0410-87a9-c151d255dfc7